backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0
sqlalchemy>=2.0
alembic
psycopg2-binary
python-telegram-bot>=21.0
httpx
python-jose[cryptography]
python-dotenv
